+++ packages/cli/src/index.ts
@@ -19,6 +19,8 @@ import { version } from './update-version-notifier';
 import type { Arguments } from 'yargs';
 import type { OutputFormat, RuleSeverity } from '@redocly/openapi-core';
 import type { BuildDocsArgv } from './commands/build-docs/types';
+import { previewProject } from './commands/preview-project';
+import { PRODUCT_PLANS } from './commands/preview-project/constants';
 
 if (!('replaceAll' in String.prototype)) {
   require('core-js/actual/string/replace-all');
@@ -92,7 +94,7 @@ yargs
   )
   .command(
     'join [apis...]',
-    'Join definitions [experimental].',
+    'Join multiple API descriptions into one [experimental].',
     (yargs) =>
       yargs
         .positional('apis', {
@@ -101,7 +103,7 @@ yargs
           demandOption: true,
         })
         .option({
-          lint: { description: 'Lint definitions', type: 'boolean', default: false, hidden: true },
+          lint: { description: 'Lint descriptions', type: 'boolean', default: false, hidden: true },
           decorate: { description: 'Run decorators', type: 'boolean', default: false },
           preprocess: { description: 'Run preprocessors', type: 'boolean', default: false },
           'prefix-tags-with-info-prop': {
@@ -124,7 +126,7 @@ yargs
             type: 'boolean',
           },
           output: {
-            describe: 'Output file',
+            description: 'Output file.',
             alias: 'o',
             type: 'string',
           },
@@ -228,7 +230,7 @@ yargs
   )
   .command(
     'lint [apis...]',
-    'Lint definition.',
+    'Lint an API description.',
     (yargs) =>
       yargs.positional('apis', { array: true, type: 'string', demandOption: true }).option({
         format: {
@@ -287,10 +289,14 @@ yargs
   )
   .command(
     'bundle [apis...]',
-    'Bundle definition.',
+    'Bundle a multi-file API description to a single file.',
     (yargs) =>
       yargs.positional('apis', { array: true, type: 'string', demandOption: true }).options({
-        output: { type: 'string', alias: 'o' },
+        output: {
+          type: 'string',
+          description: 'Output file.',
+          alias: 'o',
+        },
         format: {
           description: 'Use a specific output format.',
           choices: ['stylish', 'codeframe', 'json', 'checkstyle'] as ReadonlyArray<OutputFormat>,
@@ -414,8 +420,40 @@ yargs
     }
   )
   .command(
+    'preview',
+    'Preview Redocly project using one of the product NPM packages.',
+    (yargs) =>
+      yargs.options({
+        product: {
+          type: 'string',
+          choices: ['redoc', 'revel', 'reef', 'realm', 'redoc-revel', 'redoc-reef', 'revel-reef'],
+          description:
+            "Product used to launch preview. Default is inferred from project's package.json or 'realm' is used.",
+        },
+        plan: {
+          type: 'string',
+          choices: PRODUCT_PLANS,
+          default: 'enterprise',
+        },
+        port: {
+          type: 'number',
+          description: 'Preview port.',
+          default: 4000,
+        },
+        'source-dir': {
+          alias: 'd',
+          type: 'string',
+          description: 'Project directory.',
+          default: '.',
+        },
+      }),
+    (argv) => {
+      commandWrapper(previewProject)(argv);
+    }
+  )
+  .command(
     'preview-docs [api]',
-    'Preview API reference docs for the specified definition.',
+    'Preview API reference docs for an API description.',
     (yargs) =>
       yargs.positional('api', { type: 'string' }).options({
         port: {
@@ -519,7 +557,7 @@ yargs
       commandWrapper(handlerBuildCommand)(argv as Arguments<BuildDocsArgv>);
     }
   )
-  .completion('completion', 'Generate completion script.')
+  .completion('completion', 'Generate autocomplete script for `redocly` command.')
   .demandCommand(1)
   .middleware([notifyUpdateCliVersion])
   .strict().argv;